moving elastography image container class to elastography module
